add DeadlineExpiredException
Adds a base exception type DeadlineExpiredException, as well as two
subtypes External and Internal.

A future change will modify the library to throw one of these exception
types when a deadline is exceeded, but for now the types are added to
the public API to allow consumers to adopt logic to handle each
exception type gracefully if necessary.