dont report expired metrics when storing deadlines
We're potentially double-counting the expired deadline meter, as we
currently call `checkExpiration` from both `encodeToRequest` and
`parseFromRequest`. Because there is no enforcement currently, this
mostly just serves to have services that continue RPC hops mark the
meter twice: first when an `Expect-Within: 0` is received on the wire,
and again when another outbound request is made.

Checking for expiration on the store path isn't really necessary
currently; we can add this back in when we enable enforcement, just to
ensure that if by chance an expired deadline reaches a service via the
header, we fail immediately (this could happen if something encodes an
expired deadline header outside of the `encodeToRequest` codepath within
this library, though this is hopefully unlikely).