allow negative deadlines returned by getRemainingDeadline
Previously we always returned Duration.ZERO if the deadline had expired.
This is straightforward and keeps some detail hidden from consumers, at
the expense of some flexibility.

With this change, just return the raw remaining deadline value as a
Duration, which may be less than zero. This allows consumers the
flexibility to report observability metrics on deadline expiration (for
instance, they can know how far past a deadline some call chain has gone
which may be of interest in debugging).