add ability to observe deadline expiration counts
Provides the ability to observe the number of deadline expirations that
happen between two points in time. Intended for consumers to be able
to potentially build more customized observability outside of the
metrics reported directly by this library.

A `ExpirationObservation` just reads the value of the internal meters
between two points in time and exposes a count of the number of
expirations that occurred.