require initialization of deadline handling
Previously, call to encodeToRequest() would _always_ encode a deadline
header on the request, even if no deadline state had been populated by a
previous call to parseFromRequest. This is generally fine, but there are
potential workflows where we might not want to do that such as
background tasks that run outside the confines of handling a request.

With this change, deadline handling must be initialized first for the
current trace, either explicitly by calling initializeForCurrentTrace()
or implicitly via a call to parseFromRequest() that stores deadline
state. If deadline handling remains uninitialized, then calls to
encodeToRequest() will result in a no-op and no deadlines will be set on
outbound requests until handling is initialized via one of the methods
described above.