deadline.expired metric reports a "throw" intent
This is distinguished from "propagate", and indicates that an exception
will be thrown because enforcement was enabled (whereas "propagate"
marks the meter to indicate a deadline expiration, but does _not_ throw
an exception).

This is useful for distinguishing codepaths that hit a deadline
expiration causing the caller to receive an exception, versus ones that
allow `encodeToRequest` to succeed while only reporting a metric
indicating that the deadline had expired.